webapk: Fix signed vs unsigned comparison warning
shell-encryption and private_membership leaked some configs
which disable some default warnings.

In order to re-enable all these warnings for chrome/browser/*
in this CL crrev.com/c/3303510, all warnings need to be fixed.

This CL is fixing the comparison between two integers for them
to be from the same exact type (i.e. integer).

Bug: chromium:1212837, b:201763017
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3320445
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>
Cr-Commit-Position: refs/heads/main@{#956347}

+++ chrome/browser/apps/app_service/webapk/webapk_install_task.cc
@@ -132,7 +132,8 @@ bool DoesShareTargetDiffer(webapk::WebAppManifest manifest,
   }
 
   // Compare share files.
-  if (share_param.files_size() != share_info->file_names.size()) {
+  if (share_param.files_size() !=
+      static_cast<int>(share_info->file_names.size())) {
     return true;
   }
 
@@ -142,7 +143,7 @@ bool DoesShareTargetDiffer(webapk::WebAppManifest manifest,
     }
 
     if (share_param.files(i).accept_size() !=
-        share_info->file_accepts[i].size()) {
+        static_cast<int>(share_info->file_accepts[i].size())) {
       return true;
     }
